refactor(autocapture): move walkUpToClickableParent from ClickOptions to AutocaptureOptions
The flag affects $el_id resolution for all autocapture event types
($mp_click, $mp_rage_click, $mp_dead_click), not just clicks. Moving
it to AutocaptureOptions reflects that it is a cross-cutting element
identification concern rather than a click-specific setting.
