+++ analytics/src/main/java/com/mixpanel/android/mpmetrics/ClickOptions.java
@@ -20,11 +20,9 @@ package com.mixpanel.android.mpmetrics;
 public class ClickOptions {
 
     private final boolean mEnabled;
-    private final boolean mWalkUpToClickableParent;
 
     private ClickOptions(Builder builder) {
         this.mEnabled = builder.mEnabled;
-        this.mWalkUpToClickableParent = builder.mWalkUpToClickableParent;
     }
 
     /**
@@ -38,16 +36,6 @@ public class ClickOptions {
     }
 
     /**
-     * Returns whether walk-up-to-clickable-parent is enabled.
-     *
-     * @return {@code true} if the SDK should walk up to the nearest clickable ancestor
-     *         when the tapped view has no meaningful identity.
-     */
-    public boolean isWalkUpToClickableParent() {
-        return mWalkUpToClickableParent;
-    }
-
-    /**
      * Builder for creating {@link ClickOptions} instances.
      *
      * <p>Default values:
@@ -57,7 +45,6 @@ public class ClickOptions {
      */
     public static class Builder {
         private boolean mEnabled = true;
-        private boolean mWalkUpToClickableParent = false;
 
         public Builder() {
         }
@@ -69,7 +56,6 @@ public class ClickOptions {
          */
         public Builder(ClickOptions source) {
             this.mEnabled = source.mEnabled;
-            this.mWalkUpToClickableParent = source.mWalkUpToClickableParent;
         }
 
         /**
@@ -80,29 +66,6 @@ public class ClickOptions {
          */
         public Builder enabled(boolean enabled) {
             this.mEnabled = enabled;
-            return this;
-        }
-
-        /**
-         * When enabled, if the tapped view has no meaningful identifier
-         * (contentDescription or valid resource ID), the SDK walks up the
-         * view hierarchy to the nearest clickable ancestor and uses its
-         * identity instead.
-         *
-         * <p>This is intended for cross-platform frameworks (e.g., React Native,
-         * Flutter) where interactive components are wrappers around leaf views
-         * that don't carry their own identity. It only affects the
-         * {@code android.view.View} extraction path — Compose semantics
-         * extraction is not affected.
-         *
-         * <p>Defaults to {@code false}. Native Android apps should generally
-         * leave this disabled.
-         *
-         * @param walkUp {@code true} to enable ancestor walk-up
-         * @return This Builder instance for chaining.
-         */
-        public Builder walkUpToClickableParent(boolean walkUp) {
-            this.mWalkUpToClickableParent = walkUp;
             return this;
         }
